backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.23
pydantic==2.7.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
